Improved event handling via extending AggregateRoot<T>. Added entity tests. Improved entity methods.

+++ src/main/java/com/marwan/ecommerce/service/purchase/command/CreatePurchaseCommand.java
@@ -1,10 +1,11 @@
 package com.marwan.ecommerce.service.purchase.command;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CreatePurchaseCommand(
         UUID productId,
-        double unitPrice,
+        BigDecimal unitPrice,
         int quantity,
         UUID supplierId
 )
